# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GDSC Test API",
        "description": "API documentation for the GDSC Test collection",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://joe-gdsc-8be20852b361.herokuapp.com/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Read Books",
                "description": "Retrieves all book entries.",
                "operationId": "ba27de62d1f8d5416721c5155065630c",
                "responses": {
                    "200": {
                        "description": "List of books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "author": {
                                                        "type": "string"
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create Book",
                "description": "Creates a new book entry.",
                "operationId": "85da5e21c90d94e519ef3df0ecf457c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "published_at"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "5 Minute to Learn Go"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "Sundar Pichai"
                                    },
                                    "published_at": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "5 Minute to Learn Go"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Sundar Pichai"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{bookId}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get Book",
                "description": "Retrieves a specific book by ID.",
                "operationId": "11dca17273b7abe35c33572f49b9c1d0",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update Book",
                "description": "Updates the details of a specific book.",
                "operationId": "641785902f2266476a933699c165f629",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Updated Title"
                                                },
                                                "author": {
                                                    "type": "string"
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete Book",
                "description": "Deletes a specific book.",
                "operationId": "e363655805acd7689906fe86eb7f80af",
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "API operations related to books"
        }
    ]
}